Adding full page scroll
There's A LOT of clean up that needs to be done. This is my first time
actually looking through the code. I think we need to go over everything
together over a video call or something. I've changed a lot and there
are some errors somewhere causing certain parts to look weird. But I
decided to push nonetheless because we need to start coding this website
properly. We can debug together.